+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 #### 功能构建
 
+- 导入运维指令。
+  - com.dwarfeng.datamark.service.telqos.*。
+
+- 增加 Hibernate 实体数据标记字段，并应用相关实体侦听器。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateDispatcherInfo。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateMetaIndicator。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateNotifySetting。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateRouterInfo。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateSenderInfo。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateTopic。
+  - com.dwarfeng.notify.impl.bean.entity.HibernateUser。
+
+- 增加依赖。
+  - 增加依赖 `dwarfeng-datamark` 以应用其新功能，版本为 `1.0.1.a`。
+
 - 优化部分单元测试代码，以规避潜在的 bug。
   - com.dwarfeng.notify.impl.service.DispatcherInfoMaintainServiceImplTest。
   - com.dwarfeng.notify.impl.service.DispatcherSupportMaintainServiceImplTest。